nfc: move serialization interface to `HLSLRootSignature`
